requirements.txt
cupy-cuda13x==13.6.0
fastrlock==0.8.3
numpy==2.3.5
matplotlib==3.10.7
hypothesis==6.136.9
ase==3.26.0
